+++ RELEASE_NOTES.md
@@ -5,6 +5,12 @@ Version 0.3.0 (incoming)
  * 3D vectors
  * documentation updates
  
+Version 0.2.8 (March 2nd 2017)
+------------------------------
+ * Added multiplyScaledVectorBy() and divideScaledVectorBy() methods to the Vector2D class
+ * Added maxTo() and minTo() methods to the Vector2D class
+ * Added maxScalarTo() and minScalarTo() to the Vector2D class
+ 
 Version 0.2.7 (February 26th 2017)
 ------------------------------
  * Added strokeColor and strokeWidth parameters to the draw() method of the Geometry.Circle class